Use same server file for all E2E tests (#2165)
