+++ .github/workflows/e2e.yml
@@ -86,13 +86,13 @@ jobs:
       - name: Install dependencies
         run: python -m poetry install
       - name: Download dataset
-        if: ${{matrix.dataset}}
+        if: ${{ matrix.dataset }}
         run: python -c "${{ matrix.dataset }}"
       - name: Run edge client test
-        run: ./../test.sh
-      - name: Run virtual client test 
+        run: ./../test.sh "${{ matrix.directory }}"
+      - name: Run virtual client test
         run: python simulation.py
-      - name: Run driver test 
+      - name: Run driver test
         run: ./../test_driver.sh
 
   strategies:
