+++ e2e/bare/client.py
@@ -23,6 +23,10 @@ class FlowerClient(fl.client.NumPyClient):
         accuracy = 1 - loss
         return loss, 1, {"accuracy": accuracy}
 
+def client_fn(cid):
+    return FlowerClient()
+
+
 if __name__ == "__main__":
     # Start Flower client
     fl.client.start_numpy_client(server_address="127.0.0.1:8080", client=FlowerClient())
