+++ e2e/fastai/client.py
@@ -49,6 +49,10 @@ class FlowerClient(fl.client.NumPyClient):
         return loss, len(dls.valid), {"accuracy": 1 - error_rate}
 
 
+def client_fn(cid):
+    return FlowerClient()
+
+
 if __name__ == "__main__":
     # Start Flower client
     fl.client.start_numpy_client(
