+++ e2e/jax/client.py
@@ -50,6 +50,9 @@ class FlowerClient(fl.client.NumPyClient):
         )
         return float(loss), num_examples, {"loss": float(loss)}
 
+def client_fn(cid):
+    return FlowerClient()
+
 if __name__ == "__main__":
     # Start Flower client
     fl.client.start_numpy_client(server_address="127.0.0.1:8080", client=FlowerClient())
