+++ e2e/pandas/client.py
@@ -33,6 +33,9 @@ class FlowerClient(fl.client.NumPyClient):
             {},
         )
 
+def client_fn(cid):
+    return FlowerClient()
+
 if __name__ == "__main__":
     # Start Flower client
     fl.client.start_numpy_client(
