+++ e2e/pandas/server.py
@@ -2,12 +2,12 @@ import flwr as fl
 
 from strategy import FedAnalytics
 
-# Start Flower server
 hist = fl.server.start_server(
     server_address="0.0.0.0:8080",
     config=fl.server.ServerConfig(num_rounds=1),
     strategy=FedAnalytics(),
 )
+
 assert hist.metrics_centralized["Aggregated histograms"][1][1] == [
     "Length:",
     "18",
