--- e2e/pandas/test.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-set -e
-cd "$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"/
-
-bash ../test.sh
