+++ e2e/tabnet/client.py
@@ -78,6 +78,9 @@ class FlowerClient(fl.client.NumPyClient):
         return loss, len(ds_train), {"accuracy": accuracy}
 
 
+def client_fn(cid):
+    return FlowerClient()
+
 if __name__ == "__main__":
     # Start Flower client
     fl.client.start_numpy_client(server_address="127.0.0.1:8080", client=FlowerClient())
