--- e2e/tabnet/server.py
@@ -1,7 +1,0 @@
-import flwr as fl
-
-hist = fl.server.start_server(
-    server_address="0.0.0.0:8080",
-    config=fl.server.ServerConfig(num_rounds=3),
-)
-assert (hist.losses_distributed[0][1] / hist.losses_distributed[-1][1]) >= 1
