+++ e2e/tensorflow/client.py
@@ -31,6 +31,8 @@ class FlowerClient(fl.client.NumPyClient):
         loss, accuracy = model.evaluate(x_test, y_test)
         return loss, len(x_test), {"accuracy": accuracy}
 
+def client_fn(cid):
+    return FlowerClient()
 
 if __name__ == "__main__":
     # Start Flower client
